+++ lib/rubocop/cop/rspec/subject_stub.rb
@@ -18,7 +18,8 @@ module RuboCop
       #   end
       #
       class SubjectStub < Cop
-        include RuboCop::RSpec::TopLevelDescribe,
+        include RuboCop::RSpec::SpecOnly,
+                RuboCop::RSpec::TopLevelDescribe,
                 RuboCop::RSpec::Language,
                 RuboCop::RSpec::Language::NodePattern