+++ spec/rubocop/cop/rspec/example_wording_spec.rb
@@ -1,6 +1,8 @@
 describe RuboCop::Cop::RSpec::ExampleWording, :config do
   subject(:cop) { described_class.new(config) }
 
+  include_examples 'an rspec only cop'
+
   context 'with configuration' do
     let(:cop_config) do
       {
@@ -45,19 +47,37 @@ describe RuboCop::Cop::RSpec::ExampleWording, :config do
     end
 
     it 'corrects `it "should only have"` to it "only has"' do
-      corrected = autocorrect_source(cop, 'it "should only have trait" do end')
+      corrected =
+        autocorrect_source(
+          cop,
+          'it "should only have trait" do end',
+          'spec/foo_spec.rb'
+        )
+
       expect(corrected).to eql('it "only has trait" do end')
     end
   end
 
   context 'when configuration is empty' do
     it 'only does not correct "have"' do
-      corrected = autocorrect_source(cop, 'it "should have trait" do end')
+      corrected =
+        autocorrect_source(
+          cop,
+          'it "should have trait" do end',
+          'spec/foo_spec.rb'
+        )
+
       expect(corrected).to eql('it "haves trait" do end')
     end
 
     it 'only does not make an exception for the word "only"' do
-      corrected = autocorrect_source(cop, 'it "should only fail" do end')
+      corrected =
+        autocorrect_source(
+          cop,
+          'it "should only fail" do end',
+          'spec/foo_spec.rb'
+        )
+
       expect(corrected).to eql('it "onlies fail" do end')
     end
   end